Update addon and english locales to prerelease
