+++ addon/DomHelpers.js
@@ -1,0 +1,105 @@
+export default class DomHelpers {
+  constructor(addon) {
+    this.addon = addon;
+    this.vm = addon.tab.traps.vm;
+    /**
+     * @type {eventDetails[]}
+     */
+    this.events = [];
+  }
+
+  /**
+   * Simulate a drag and drop programmatically through javascript
+   * @param selectorDrag
+   * @param selectorDrop
+   * @param mouseXY
+   * @returns {boolean}
+   */
+  triggerDragAndDrop(selectorDrag, selectorDrop, mouseXY) {
+    // function for triggering mouse events
+    let fireMouseEvent = function (type, elem, centerX, centerY) {
+      let evt = document.createEvent("MouseEvents");
+      evt.initMouseEvent(type, true, true, window, 1, 1, 1, centerX, centerY, false, false, false, false, 0, elem);
+      elem.dispatchEvent(evt);
+    };
+
+    // fetch target elements
+    let elemDrag = selectorDrag; // document.querySelector(selectorDrag);
+    let elemDrop = selectorDrop; // document.querySelector(selectorDrop);
+    if (!elemDrag /* || !elemDrop*/) {
+      return false;
+    }
+
+    // calculate positions
+    let pos = elemDrag.getBoundingClientRect();
+    let center1X = Math.floor((pos.left + pos.right) / 2);
+    let center1Y = Math.floor((pos.top + pos.bottom) / 2);
+
+    // mouse over dragged element and mousedown
+    fireMouseEvent("mouseover", elemDrag, center1X, center1Y);
+    fireMouseEvent("mousedown", elemDrag, center1X, center1Y);
+
+    // start dragging process over to drop target
+    fireMouseEvent("dragstart", elemDrag, center1X, center1Y);
+    fireMouseEvent("drag", elemDrag, center1X, center1Y);
+    fireMouseEvent("mousemove", elemDrag, center1X, center1Y);
+
+    if (!elemDrop) {
+      if (mouseXY) {
+        // console.log(mouseXY);
+        let center2X = mouseXY.x;
+        let center2Y = mouseXY.y;
+        fireMouseEvent("drag", elemDrag, center2X, center2Y);
+        fireMouseEvent("mousemove", elemDrag, center2X, center2Y);
+      }
+      return false;
+    }
+
+    pos = elemDrop.getBoundingClientRect();
+    let center2X = Math.floor((pos.left + pos.right) / 2);
+    let center2Y = Math.floor((pos.top + pos.bottom) / 2);
+
+    fireMouseEvent("drag", elemDrag, center2X, center2Y);
+    fireMouseEvent("mousemove", elemDrop, center2X, center2Y);
+
+    // trigger dragging process on top of drop target
+    fireMouseEvent("mouseenter", elemDrop, center2X, center2Y);
+    fireMouseEvent("dragenter", elemDrop, center2X, center2Y);
+    fireMouseEvent("mouseover", elemDrop, center2X, center2Y);
+    fireMouseEvent("dragover", elemDrop, center2X, center2Y);
+
+    // release dragged element on top of drop target
+    fireMouseEvent("drop", elemDrop, center2X, center2Y);
+    fireMouseEvent("dragend", elemDrag, center2X, center2Y);
+    fireMouseEvent("mouseup", elemDrag, center2X, center2Y);
+
+    return true;
+  }
+
+  bindOnce(dom, event, func, capture) {
+    capture = !!capture;
+    dom.removeEventListener(event, func, capture);
+    dom.addEventListener(event, func, capture);
+    this.events.push(new eventDetails(dom, event, func, capture));
+  }
+
+  unbindAllEvents() {
+    console.log("Unbinding Events - gui has become dirty");
+    for (const event of this.events) {
+      event.dom.removeEventListener(event.event, event.func, event.capture);
+    }
+    this.events = [];
+  }
+}
+
+/**
+ * A record of an event
+ */
+class eventDetails {
+  constructor(dom, event, func, capture) {
+    this.dom = dom;
+    this.event = event;
+    this.func = func;
+    this.capture = capture;
+  }
+}
