+++ addon/blockly/UndoGroup.js
@@ -1,0 +1,47 @@
+/**
+ * This class is dedicated to maintaining the Undo stack of Blockly
+ * It allows us to initiate an undo group such that all subsequent operations are recorded as a single
+ * undoable transaction.
+ */
+export default class UndoGroup {
+  /**
+   * Start an Undo group - begin recording
+   * @param workspace the workspace
+   */
+  static startUndoGroup(workspace) {
+    const undoStack = workspace.undoStack_;
+    if (undoStack.length) {
+      undoStack[undoStack.length - 1]._devtoolsLastUndo = true;
+    }
+  }
+
+  /**
+   * End an Undo group - stops recording
+   * @param workspace the workspace
+   */
+  static endUndoGroup(workspace) {
+    const undoStack = workspace.undoStack_;
+    // Events (responsible for undoStack updates) are delayed with a setTimeout(f, 0)
+    // https://github.com/LLK/scratch-blocks/blob/f159a1779e5391b502d374fb2fdd0cb5ca43d6a2/core/events.js#L182
+    setTimeout(() => {
+      const group = generateUID();
+      for (let i = undoStack.length - 1; i >= 0 && !undoStack[i]._devtoolsLastUndo; i--) {
+        undoStack[i].group = group;
+      }
+    }, 0);
+  }
+}
+
+/**
+ * https://github.com/LLK/scratch-blocks/blob/f159a1779e5391b502d374fb2fdd0cb5ca43d6a2/core/events.js#L182
+ * @returns {string}
+ * @private
+ */
+function generateUID() {
+  const CHARACTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!#$%()*+,-./:;=?@[]^_`{|}~";
+  let result = "";
+  for (let i = 0; i < 20; i++) {
+    result += CHARACTERS[Math.floor(Math.random() * CHARACTERS.length)];
+  }
+  return result;
+}
