+++ addon/userscript.css
@@ -227,7 +227,7 @@ div.s3devDDOut.vis ul.s3devDD {
 }
 .data-lists:hover,
 .data-lists.sel {
-  background-color: rgb(255, 102, 26);
+  background-color: rgb(207, 81, 20);
 }
 
 .sensing {
